Add files via upload
Phase 1 and Phase 2 of the assignment will result in a mobile, distributed system called MonashFriendFinder which helps students search new university friends. This report focuses on the phase 1, it is responsible for the RESTful web services in NetBeans in Java where the data and business logic make up backend of the system. It contains some details of the developed RESTful service.